the project as a .zip
I am hitting every wall trying to get this to work and this is the best I can do. I either don't have permission, it doesn't see the repo, or hits a 403 error and every solution on the internet isn't working. hopefully, this gets the file to you. I will look this weekend into trying to understand github more.